Improved physics Player with crate
- Added physics material to both the crate and player and set the friction to 0. This prevents the player from slowly hovering/climbing up the crates.
- Added in a check to see if a crate is landing on top of another crate, if so it bounces the colliding crate back up.

NOTE:
Add in a check if the player hits the crate from below while it's falling down to see if the crate is either a breakable or interactable. If it's breakable it should break according to to the crate type. If it's interactable it means it is a metal crate so the player should die from it.