Nitro crate added
- Added in nitro crate explosion effect when the nitro crate explodes.
- Nitro crate explodes when it gets hit by an AOE from another explosion, enemy or player.
- Nitro crate explodes regardless how the player interacts with it.

TO DO:
Build in check to see if the player has either invinsibility or an extra hit. If the player has invinsibility only the crate needs to explode the crate and not hurt the player. If the player has an extra hit the crate needs to explode and has to remove that extra hit point from the player and give the player a couple invinsibility frames to recover.