Bug fix Activator script
Issue was that I would get an error back after the player broke a couple of crates from the activator, killed himself, then activated it again and tried to break them again.

Fixed by adding in a temporarely FilterList that keeps track of all the crates that haven't been destroyed yet and sets the CratesList equal to that list once the game needs to be reset after the players death.