Level Manager implemented
- Ripped out the player reference, fade effect  and game event raiser from the BoxCounter script and implemented this into the Level Manager.

This way the BoxCounter only has to check if all the crates are broken to spawn in the gem or not. This way the player can still die afterwards when collecting the gem. This way I can also check the players life count so I can check if the level needs to be reset or that the game over needs to be displayed.