Improved exclamation mark animation
Looked back at the orginal Crash Bandicoot and saw that they also had an exclamation mark on top of the crate and this one lasted longer the other once on the rest of the sides. I've added this in as well and the improvement looks a lot better and smoother now.