Game over scene added
Added in a check to see if the players lives are below 0. If not it resets the player position and crate count. If it is the game over scene will be displayed and the scene will be reloaded after 5 seconds.

TO DO:
Instead of reloading the scene the players should go back to the world select scene once this has been implemented.