Fixed checkpoint clipping issue
Issue was when a player set the checkpoint at an interactable checkpoint and would respawn at that position the player would be inside the metal crate and be shot out of it to prevent getting stuck.

Solved by, when setting the players checkpointposition to that checkpoint position I add an extra height to it so the player would respawn on top of the crate instead of inside it to prevent it from clipping.