Improved Checkpoint reset
Instead of adjusting the players height position when the new position has already been passed throught I check when the player hits an interactable checkpoint to add the extra height at that position. Otherwise it's was a breakable checkpoint crate which means the player can just spawn at it's origin position.