Updated Level Manager script
Added in an extra check to see when the gems collection has a box gem it checks if the ID matches with the available gem ID that is also a white box gem. If it is that means the player has already collected the box gem for that level and the total crate counter will now get spawned in.

TO DO:
Fix issue where colored gem keeps getting spawned in while already been collected.