New Gem Collection system implemented
The last version of the gem collection system didn't work the way I wanted to. I've worked on a new system with another Game Developer and the new system works as intended. Once the player exits the level the progress gets saved with the current amount of gems collected, live, Aku-Aku and Woompa count. Once the player loads in a level it checks if the player has already collected the gem in that particular level it disables the spawn color gem locations and crate counter if the player already has collected the box gem from that level.