Temporarely invulnerability added
Once the player has been hit by an enemie and he has an extra hit left, this will be removed from the player and he will get temporarely invulnerability frames so the player has a fair change of escaping.

TO DO:
Once the player hit an enemy this enemy will then get killed of if this is allowed. Else the enemy will just stay alive but won't be able to attack the player until it's allowed again.