Extra hit mechanic setup
Added in the visuals to represent what state you're current Extra is in. When a player gets hit the visual representation will get updated. If the player has three extra hits the visual will be positioned in front of the players face and can't take any damage unless it's with an instakill object.
Made it so that when the player dies from an instakill object the player loses all the masks he currently had on him, even if he was in invinsability mode.

TO DO:
- Add Invinsibility effect to the player when he has collected three extra hits.
- Remove third extra hit after a certain amount of time is passed and reposition the the remaining extra hits back to the side of the player.